AUTH-SERVICE/requirements.txt
fastapi[standard]==0.121.1
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6